output type

+++ cs336_systems/flash_attention/flash_attention_triton.py
@@ -309,8 +309,8 @@ class FlashAttention(torch.autograd.Function):
         d = Q.shape[2]
 
         dQ = torch.empty((batch_size, N_q, d), device=Q.device, dtype=Q.dtype)
-        dK = torch.zeros((batch_size, N_k, d), device=K.device, dtype=K.dtype)
-        dV = torch.zeros((batch_size, N_k, d), device=V.device, dtype=V.dtype)
+        dK = torch.zeros((batch_size, N_k, d), device=K.device, dtype=torch.float32)
+        dV = torch.zeros((batch_size, N_k, d), device=V.device, dtype=torch.float32)
 
         flash_bwd_kernel[(triton.cdiv(N_q, B_q), batch_size)](
             Q, K, V, O, dO, L,
@@ -332,4 +332,4 @@ class FlashAttention(torch.autograd.Function):
             ctx.is_causal,
         )
 
-        return dQ, dK, dV, None
+        return dQ, dK.to(K.dtype), dV.to(V.dtype), None
